updated skills list w/ cont edu goals
Updated the list items for my skills to include the skills I am currently learning which is everything cloud. I would like to be cloud agnostic with the ability to work in any cloud. My learning style is always the big picture whereas the details are all resources that I know where to find.